[Refactor]: 공용 kakaoMap 구현에 따른 별도의 SaveDetailMap 구현 및 불필요한 파일 삭제
